addCardToInventory moved to interactors
